+++ app/assets/Data/books.js
@@ -1,20 +1,20 @@
 const BOOKS = [
     {
-        id: 'bd7acbea-c1b1-46c2-aed5-3ad53abb28bb',
-        name: 'Rafael Gómez Blanes',
-        book: 'El Libro Negro del Programador',
+        _id: 'bd7acbea-c1b1-46c2-aed5-3ad53abb28bb',
+        author: 'Rafael Gómez Blanes',
+        title: 'El Libro Negro del Programador',
         link: 'https://www.amazon.com/Libro-Negro-del-Programador-desarrollando-ebook/dp/B00J5RHHNW',
     },
     {
-        id: 'bd7acbea-c1b1-46c2-aed5-3ad53abb28bc',
-        name: 'Robert C. Martin',
-        book: 'El limpiador de código: Código de conducta para programadores profesionales',
+        _id: 'bd7acbea-c1b1-46c2-aed5-3ad53abb28bc',
+        author: 'Robert C. Martin',
+        title: 'El limpiador de código: Código de conducta para programadores profesionales',
         link: 'https://www.amazon.com/-/es/Robert-C-Martin/dp/8441540845',
     },
     {
-        id: 'bd7acbea-c1b1-46c2-aed5-3ad53abb28ba',
-        name: 'Phil Knight',
-        book: 'Nunca pares: Autobiograf\u00eda del fundador de Nike',
+        _id: 'bd7acbea-c1b1-46c2-aed5-3ad53abb28ba',
+        author: 'Phil Knight',
+        title: 'Nunca pares: Autobiograf\u00eda del fundador de Nike',
         link: 'https://www.amazon.com/Nunca-pares-Autobiograf%C3%ADa-fundador-Creator/dp/6073149824',
     },
 ];