+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+#v0.3.0
+
+- [feat] Consume the books endpoint.
+
 #v0.2.0
 
 - [feat] Add book screen.